Move similar files to new questions directory
