+++ client/js/mini-app/questions/resource-card.js
